refactor(#1) categoryList 적용
