+++ src/users/entities/user-badge.entity.ts
@@ -6,7 +6,7 @@ import {
   PrimaryGeneratedColumn,
 } from 'typeorm';
 import { User } from './user.entity';
-import { BadgeList } from './badge-list.entity';
+import { BadgeList } from '../../common/entity/badge-list.entity';
 
 @Entity({ name: 'user_badge' })
 export class UserBadge {
